Using BigInteger instead of int

+++ src/main/java/model/CSMT.java
@@ -3,8 +3,10 @@ package model;
 import model.proof.Proof;
 import org.jetbrains.annotations.NotNull;
 
+import java.math.BigInteger;
+
 public interface CSMT<V, H> {
-    void insert(int key, @NotNull V value);
-    void remove(int key);
-    @NotNull Proof<V, H> getProof(int key);
+    void insert(BigInteger key, @NotNull V value);
+    void remove(BigInteger key);
+    @NotNull Proof<V, H> getProof(BigInteger key);
 }
